requirements.txt
matplotlib==3.5.3
numpy==1.22.4
object_detection==0.1
Pillow==9.2.0
tensorflow==2.8.2
